rename test "screen" folder to "screens"
