+++ app/src/androidTest/java/com/android/partagix/ExampleInstrumentedTest.kt
@@ -2,7 +2,7 @@ package com.android.partagix
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.partagix.screen.MainScreen
+import com.android.partagix.screens.MainScreen
 import com.android.partagix.ui.MainActivity
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.github.kakaocup.compose.node.element.ComposeScreen
