+++ app/src/test/java/com/android/partagix/ExampleRobolectricTest.kt
@@ -2,7 +2,7 @@ package com.android.partagix
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.partagix.screen.SecondScreen
+import com.android.partagix.screens.SecondScreen
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import org.junit.Rule
