+++ app/src/test/java/com/android/partagix/screens/SecondScreen.kt
@@ -1,4 +1,4 @@
-package com.android.partagix.screen
+package com.android.partagix.screens
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
 import com.android.partagix.resources.C
